Add port to devStart.js example
I tried using this example to test a driver I just started working on and was puzzled about the INVALID_STARTSERVER_PARAMETER exception I was getting until I looked at the function throwing the exception and noticed that the port parameter is required. The port is also shown as required in the [SDK documentation]( https://neeoinc.github.io/neeo-sdk/#lib-index.js-module.exports.startserver), so this is the only place I found this error.